Create HolidayTable.java
HolidayTable is a class used to access holiday data from the database. Functions that are implemented include deterring if a certain holiday has been entered into the table yet, getting the number of holidays currently in the data table, and getting the names of the holidays currently in the table.